fix show prepayment discount on booking detail

+++ database/migrations/2022_03_11_202154_add_prepayment_discount_bookings_table.php
@@ -14,7 +14,7 @@ class AddPrepaymentDiscountBookingsTable extends Migration
     public function up()
     {
         Schema::table('bookings', function (Blueprint $table) {
-            $table->float('prepayment_discount_percent')->after('discount_percent');
+            $table->integer('prepayment_discount_percent')->after('discount_percent');
         });
     }
